LayoutTests/http/tests/security/cross-frame-access-name-getter-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-name-getter.html. Domains, protocols and ports must match.



----- tests for getting properties by name -----

PASS: canGet('targetWindow.testiframe') should be 'true' and is.
PASS: canGet('targetWindow.testimage') should be 'false' and is.
